File upload and download, pdf generation and streaming of file to response

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require("../models/product");
 const { validationResult } = require('express-validator');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -14,9 +15,24 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+
+  if(!image) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "Add Product",
+      path: "/admin/add-product",
+      editMode: false,
+      hasError: true,
+      product: { title: title, price: price, description: description },
+      errorMessage: 'Attached file is not an image.',
+      validationErrors: [],
+    });
+  }
+
+  const imageUrl = image.path;
+
   const product = new Product({
     title: title,
     price: price,
@@ -33,7 +49,7 @@ exports.postAddProduct = (req, res, next) => {
       path: "/admin/add-product",
       editMode: false,
       hasError: true,
-      product: { title: title, imageUrl: imageUrl, price: price, description: description },
+      product: { title: title, price: price, description: description },
       errorMessage: errors.array()[0].msg ?? null,
       validationErrors: errors.array(),
     });
@@ -48,7 +64,7 @@ exports.postAddProduct = (req, res, next) => {
     })
     .catch((err) => {
       console.log(err);
-      const error = new Error('Error creating new product.');
+      const error = new Error(`Error creating new product: ${err}`);
       error.status = 500;
       next(error);
     });
@@ -83,7 +99,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
 
   errors = validationResult(req);
@@ -94,7 +110,7 @@ exports.postEditProduct = (req, res, next) => {
       path: "/admin/edit-product",
       editMode: true,
       hasError: true,
-      product: { _id: prodId, title: updatedTitle, imageUrl: updatedImageUrl, price: updatedPrice, description: updatedDesc },
+      product: { _id: prodId, title: updatedTitle, price: updatedPrice, description: updatedDesc },
       errorMessage: errors.array()[0].msg ?? null,
       validationErrors: errors.array(),
     });
@@ -108,7 +124,10 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if(image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       return product.save().then((result) => {
         console.log("UPDATED PRODUCT!");
         res.redirect("/admin/products");
@@ -132,10 +151,19 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then(() => {
-      console.log("DESTROYED PRODUCT");
-      res.redirect("/admin/products");
-    })
-    .catch((err) => console.log(err));
+
+  Product.findById(prodId)
+  .then((product) => {
+    if(!product) {
+      return next(new Error('Product not found.'));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    return Product.deleteOne({ _id: prodId, userId: req.user._id });
+  })
+  .then(() => {
+    console.log("DESTROYED PRODUCT");
+    res.redirect("/admin/products");
+  })
+  .catch((err) => console.log(err));  
+
 };
